+++ build.R
@@ -7,6 +7,7 @@ devtools::install(build_vignettes = TRUE)
 # Run the Shiny demo app (this only works when working from source)
 shiny::runApp('inst/login_demo_simple/', port = 2112)
 shiny::runApp('inst/login_template/', port = 2112)
+shiny::runApp('inst/login_modal/', port = 2112)
 
 devtools::spell_check()