+++ NEWS.md
@@ -1,6 +1,8 @@
 # login 0.9.2
 
-* Fixed an issue where additional_fields were not saved in the returend reativeValues object.
+* Fixed an issue where additional_fields were not saved in the returned reactiveValues object.
+* Used different shinybusy method so it doesn't interfere with modal dialog boxes.
+* Added new demo to use modal dialogs for logging in.
 
 # login 0.9.1