+++ NEWS.md
@@ -1,3 +1,11 @@
+# login 0.9.4
+
+* Added feature to encrypt cookie values using sodium. See the cookie_password parameter on the the login_server function for more details.
+
+# login 0.9.3
+
+* First version on CRAN.
+
 # login 0.9.2
 
 * Fixed an issue where additional_fields were not saved in the returned reactiveValues object.
@@ -6,4 +14,4 @@
 
 # login 0.9.1
 
-* Initial CRAN submission.
+* Initial version on Github.